flink-project/alpaca-api/requirements.txt
alpaca-py==0.35.0
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
kafka-python==2.0.2
msgpack==1.1.0
numpy==2.2.1
pandas==2.2.3
pyarrow==18.1.0
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sseclient-py==1.8.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websockets==14.1
